continue conversion to longs

+++ src/main/java/uk/co/real_logic/agrona/MutableDirectBuffer.java
@@ -30,7 +30,7 @@ public interface MutableDirectBuffer extends DirectBuffer
      * @param length of the run of bytes to set.
      * @param value  the memory will be set to.
      */
-    void setMemory(long index, int length, byte value);
+    void setMemory(long index, long length, byte value);
 
     /**
      * Put a value to a given index.
